+++ src/utils/Helpers.mjs
@@ -25,24 +25,6 @@ export function overrideNetworks(networks, overrides){
   })
 }
 
-export function lastRestakeClassname(operator, lastExec) {
-  if (!operator || lastExec == null)
-    return;
-
-  const missedRuns = operator.missedRunCount(lastExec);
-  const warning = missedRuns > Math.min(20, operator.runsPerDay());
-  const error = missedRuns > Math.min(30, operator.runsPerDay() * 2);
-  return error ? 'text-danger' : warning ? 'text-warning' : 'text-success';
-}
-
-export function lastRestakeWarning(operator, lastExec) {
-  if (!operator || lastExec == null)
-    return;
-
-  const missedRuns = operator.missedRunCount(lastExec);
-  return missedRuns > Math.min(30, operator.runsPerDay() * 2);
-}
-
 export function buildExecMessage(grantee, messages) {
   return {
     typeUrl: "/cosmos.authz.v1beta1.MsgExec",
